flask-api/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
coverage==6.4.4
docx2txt==0.8
doubles==1.5.3
elastic-enterprise-search==7.17.0
elastic-transport==7.16.0
Flask==1.1.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
flask-restx==0.4.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.4.0
MarkupSafe==2.0.1
mock==5.0.2
nltk==3.7
openai==0.27.2
packaging==21.3
pandas==2.0.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.8
pdfminer.six==20221105
pyrsistent==0.18.1
pytest==7.1.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2022.1
requests==2.27.1
six==1.16.0
sklearn==0.0
spacy==3.4.1
SQLAlchemy==1.4.35
stemming==1.0.1
tomli==2.0.1
urllib3==1.26.9
Werkzeug==1.0.1
cloud-sql-python-connector==1.2.0
pg8000==1.29.4
mysql-connector==2.2.9
pika==1.2.0